cars_catalog/requirements.txt
asgiref==3.8.1
bleach==6.1.0
bx_django_utils==79
bx_py_utils==101
colorlog==6.8.2
Django==5.1.1
django-debug-toolbar==4.4.6
djangorestframework==3.15.2
gunicorn==23.0.0
icdiff==2.0.7
packaging==24.1
pprintpp==0.4.0
python-dotenv==1.0.1
python-stdnum==1.20
six==1.16.0
sqlparse==0.5.1
webencodings==0.5.1
